backend/requirements.txt
openai
python-dotenv
langchain-community
qdrant-client
tiktoken